pyproject.toml
[tool.poetry]
name = "daily-learning-md"
version = "0.1.0"
description = ""
authors = ["Kaushal Partani <kaushalpartani@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
requests = "^2.32.3"
groq = "^0.11.0"
pyyaml = "^6.0.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
